+++ contrib/pg_upgrade/info.c
@@ -15,6 +15,7 @@
 #include "catalog/pg_class.h"
 #include "info_gp.h"
 
+#include "greenplum/pg_upgrade_greenplum.h"
 
 static void create_rel_filename_map(const char *old_data, const char *new_data,
 						const DbInfo *old_db, const DbInfo *new_db,
