+++ contrib/pg_upgrade/pg_upgrade.h
@@ -1,3 +1,5 @@
+#ifndef PG_UPGRADE_H
+#define PG_UPGRADE_H
 /*
  *	pg_upgrade.h
  *
@@ -677,44 +679,12 @@ bool		reap_child(bool wait_for_child);
 #endif
 #define Assert(condition) ((void) (true || (condition)))
 
-/* aotable.c */
-
-void		restore_aosegment_tables(void);
-
-/* gpdb4_heap_convert.c */
-
-const char *convert_gpdb4_heap_file(const char *src, const char *dst,
-									bool has_numerics, AttInfo *atts, int natts);
-void		finish_gpdb4_page_converter(void);
-
-/* file_gp.c */
-
 void copy_distributedlog(void);
-const char * rewriteHeapPageChecksum( const char *fromfile, const char *tofile,
-					 const char *schemaName, const char *relName);
-
-/* version_gp.c */
-
-void old_GPDB4_check_for_money_data_type_usage(void);
-void old_GPDB4_check_no_free_aoseg(void);
-void check_hash_partition_usage(void);
-void new_gpdb5_0_invalidate_indexes(void);
-void new_gpdb_invalidate_bitmap_indexes(void);
-Oid *get_numeric_types(PGconn *conn);
-void old_GPDB5_check_for_unsupported_distribution_key_data_types(void);
-
-/* check_gp.c */
-
-void check_greenplum(void);
-
-/* reporting.c */
-
-void report_progress(ClusterInfo *cluster, progress_type op, char *fmt,...)
-__attribute__((format(PG_PRINTF_ATTRIBUTE, 3, 4)));
-void close_progress(void);
 
 static inline bool
 is_gpdb6(ClusterInfo *cluster)
 {
 	return GET_MAJOR_VERSION(cluster->major_version) == 904;
 }
+
+#endif /* PG_UPGRADE_H */
