+++ contrib/pg_upgrade/server.c
@@ -12,6 +12,7 @@
 #include "fe_utils/connect.h"
 #include "pg_upgrade.h"
 
+#include "greenplum/pg_upgrade_greenplum.h"
 
 static PGconn *get_db_conn(ClusterInfo *cluster, const char *db_name);
